backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
qrcode==8.0
pillow==11.1.0
passbook==1.4.0
google-auth==2.38.0
apscheduler==3.10.4
